+++ CHANGELOG.md
@@ -5,11 +5,13 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## [NEXT_RELEASE]
 ### Added
 - Verification to avoid ingress duplication when an unmanaged ingress exists
+- Adds rewrite-target annotation to createIngress if ingressClass is nginx
 
 ### Changed
 - Default nginx side-car image to allow support to lua scripts
 - Ingress spec will be the same for vhost and static IP
 - Checks if app already has ingress before trying to reserve a static IP
+- Moves ingress annotations to deploy's exposeApp
 
 ### Fixed
 - Don't create an Ingress for apps without vHosts